+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/testing/BigtableTestUtils.java
@@ -17,6 +17,8 @@
  */
 package org.apache.beam.sdk.io.gcp.testing;
 
+import static java.nio.charset.StandardCharsets.UTF_8;
+import static java.util.stream.Collectors.toList;
 import static org.apache.beam.sdk.io.gcp.bigtable.RowUtils.KEY;
 import static org.apache.beam.sdk.io.gcp.bigtable.RowUtils.LABELS;
 import static org.apache.beam.sdk.io.gcp.bigtable.RowUtils.TIMESTAMP_MICROS;
@@ -25,8 +27,15 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.containsString;
 import static org.junit.Assert.fail;
 
+import com.google.bigtable.v2.Cell;
+import com.google.bigtable.v2.Column;
+import com.google.bigtable.v2.Family;
+import com.google.protobuf.ByteString;
 import java.time.Instant;
+import java.util.List;
 import org.apache.beam.sdk.schemas.Schema;
+import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.collect.ImmutableList;
+import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.primitives.Ints;
 import org.apache.beam.vendor.guava.v26_0_jre.com.google.common.primitives.Longs;
 import org.checkerframework.checker.nullness.qual.Nullable;
 
@@ -76,6 +85,10 @@ public class BigtableTestUtils {
   public static final long NOW = Instant.now().toEpochMilli() * 1_000;
   public static final long LATER = NOW + 1_000;
 
+  public static byte[] floatToByteArray(float number) {
+    return Ints.toByteArray(Float.floatToIntBits(number));
+  }
+
   public static byte[] doubleToByteArray(double number) {
     return Longs.toByteArray(Double.doubleToLongBits(number));
   }
@@ -90,5 +103,52 @@ public class BigtableTestUtils {
     } else {
       fail();
     }
+  }
+
+  public static com.google.bigtable.v2.Row bigTableRow() {
+    List<Column> columns =
+        ImmutableList.of(
+            column("boolColumn", booleanToByteArray(true)),
+            column("doubleColumn", doubleToByteArray(5.5)),
+            column("longColumn", Longs.toByteArray(10L)),
+            column("stringColumn", "stringValue".getBytes(UTF_8)));
+    Family family = Family.newBuilder().setName("familyTest").addAllColumns(columns).build();
+    return com.google.bigtable.v2.Row.newBuilder()
+        .setKey(ByteString.copyFromUtf8("key"))
+        .addFamilies(family)
+        .build();
+  }
+
+  // There is no possibility to insert a value with fixed timestamp so we have to replace it
+  // for the testing purpose.
+  public static com.google.bigtable.v2.Row setFixedTimestamp(com.google.bigtable.v2.Row row) {
+    Family family = row.getFamilies(0);
+
+    List<Column> columnsReplaced =
+        family.getColumnsList().stream()
+            .map(
+                column -> {
+                  Cell cell = column.getCells(0);
+                  return column(
+                      column.getQualifier().toStringUtf8(), cell.getValue().toByteArray());
+                })
+            .collect(toList());
+    Family familyReplaced =
+        Family.newBuilder().setName(family.getName()).addAllColumns(columnsReplaced).build();
+    return com.google.bigtable.v2.Row.newBuilder()
+        .setKey(row.getKey())
+        .addFamilies(familyReplaced)
+        .build();
+  }
+
+  private static Column column(String qualifier, byte[] value) {
+    return Column.newBuilder()
+        .setQualifier(ByteString.copyFromUtf8(qualifier))
+        .addCells(cell(value))
+        .build();
+  }
+
+  private static Cell cell(byte[] value) {
+    return Cell.newBuilder().setValue(ByteString.copyFrom(value)).setTimestampMicros(NOW).build();
   }
 }